Remove generics from add method and add test cases
It appears that you cannot have heterogeneous slices of types
contained in a type interface, eg []testCase[calculator.Numeric],
where the type parameter is itself a type interface.

Therefore I couldn't figure out how--or if it's even possible--to
implement an array of test cases where each case could be int or
float but all are component types of my numeric type, and so should
be suitable to pass to the various calculator functions.